Feature refactoring (#18)
* updating index slicing of tactile data in LineGraph to use direct mutable methods

* moving ThreadWorker to StateMachine

* pulling in gui style changes

* updating readme header

* updating readme header

* updating readme header

* updating overview and project background

* fixing typos in Overview

* adding new gui snapshot image to readme

* updating architecture overview and diagram

* updating gui snapshot size

* fixing typos in application architecture

* updating wording in python socket section

* updating pyqt component architecture

* fixing typos in pyqt components

* updating content in repository structure

* updating table of contents links

* rewording project background

* fixing typos

* updating gui snapshot photo

* updating gui snapshot photo

* updating gui snapshot photo
