+++ src/components/LineGraph.py
@@ -19,6 +19,7 @@ Methods:
 
 from PyQt6.QtWidgets import QWidget, QVBoxLayout
 import pyqtgraph as pg
+import math
 
 class LineGraph(QWidget):
     def __init__(self):
@@ -32,8 +33,7 @@ class LineGraph(QWidget):
         self.graph.setLabel("bottom", "Tactile Data Records")
 
         self.time = list(range(10))
-
-        self.tactile_data = {
+        self.data = {
             "x": [0 for _ in range(10)],
             "y": [0 for _ in range(10)],
             "z": [0 for _ in range(10)]
@@ -41,7 +41,7 @@ class LineGraph(QWidget):
 
         self.line_x = self.graph.plot(
             self.time,
-            self.tactile_data['x'],
+            self.data['x'],
             name="Tactile Sensor",
             pen=pg.mkPen(color="#0C746A", width=3),
             symbol="o",
@@ -51,7 +51,7 @@ class LineGraph(QWidget):
 
         self.line_y = self.graph.plot(
             self.time,
-            self.tactile_data['y'],
+            self.data['y'],
             name="Tactile Sensor",
             pen=pg.mkPen(color="#AFBBAF", width=3),
             symbol="o",
@@ -61,7 +61,7 @@ class LineGraph(QWidget):
 
         self.line_z = self.graph.plot(
             self.time,
-            self.tactile_data['z'],
+            self.data['z'],
             name="Tactile Sensor",
             pen=pg.mkPen(color="#6F8695", width=3),
             symbol="o",
@@ -77,17 +77,14 @@ class LineGraph(QWidget):
         The time and data values maintain their original length by removing the first
         element and appending the new data to the end.
         """
-        self.time = self.time[1:]
+
+        del self.time[0]
         self.time.append(self.time[-1] + 1)
-            
-        self.tactile_data['x'] = self.tactile_data['x'][1:]
-        self.tactile_data['y'] = self.tactile_data['y'][1:]
-        self.tactile_data['z'] = self.tactile_data['z'][1:]
 
-        self.tactile_data['x'].append(float(data[0]))
-        self.tactile_data['y'].append(float(data[1]))
-        self.tactile_data['z'].append(float(data[2]))
-
-        self.line_x.setData(self.time, self.tactile_data['x'])
-        self.line_y.setData(self.time, self.tactile_data['y'])
-        self.line_z.setData(self.time, self.tactile_data['z'])
+        for (key, val) in zip(self.data, data):
+            del self.data[key][0]
+            self.data[key].append(float(val))
+        
+        self.line_x.setData(self.time, self.data['x'])
+        self.line_y.setData(self.time, self.data['y'])
+        self.line_z.setData(self.time, self.data['z'])
